Refactor: useAuthProvider 훅, AuthContext, AuthActionContext을 Provider로 이동
