+++ src/_provider/AuthProvider.jsx
@@ -1,11 +1,17 @@
-import {
-  useAuthProvider,
-  AuthContext,
-  AuthActionContext,
-} from "/src/_hook/use-auth";
+import { createContext, useState } from "react";
+
+export const AuthContext = createContext(false);
+export const AuthActionContext = createContext({
+  login: () => {},
+  logout: () => {},
+});
 
 export default function AuthProvider({ children }) {
-  const [isLoggedIn, authAction] = useAuthProvider();
+  const [isLoggedIn, setIsLoggedIn] = useState(true);
+  const authAction = {
+    login: () => setIsLoggedIn(true),
+    logout: () => setIsLoggedIn(false),
+  };
 
   return (
     <AuthContext.Provider value={isLoggedIn}>
